+++ docs/source/changelog.rst
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Features:
+    * :class:`~manim_videos.VideoMObject.get_frame` now respects video masks.
+    * Add :class:`~manim_videos.VideoMObject.angle`, :class:`~manim_videos.VideoMObject.get_ordered_vertices` methods.
+
+Bug fixes:
+    * Fixed VideoMObject border not properly hiding during video overlay.
+    * Fixed output video's fps to be that of manim's config.
+
 v0.1.0 (2026-03-30)
 -------------------